+++ src/trx/config/types.h
@@ -119,6 +119,9 @@ typedef struct {
         BACKGROUND_TYPE inventory_background_style;
         BACKGROUND_TYPE stats_background_style;
         BACKGROUND_TYPE pause_background_style;
+        bool inventory_fade_effects;
+        bool stats_fade_effects;
+        bool pause_fade_effects;
 
         bool enable_smooth_bars;
         BAR_LOOK bar_look;